pyproject.toml
[project]
name = "magenta-rt-jamcli"
version = "0.1.0"
description = "CLI application for Magenta RT Audio Injection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "rich>=13.0.0",
    "click>=8.0.0",
    "numpy>=1.24.0",
    "librosa>=0.10.0",
    "jax>=0.4.0",
    "toml>=0.10.2",
    "sounddevice>=0.4.0",
    "soundfile>=0.12.0",
    "resampy>=0.4.0",
    "google-cloud-storage>=2.0.0",
    "requests>=2.28.0",
    "tqdm>=4.64.0",
]

[project.optional-dependencies]
magenta-rt = [
    # These would need to be installed from the magenta-realtime repo
    # "magenta-rt @ git+https://github.com/magenta/magenta-realtime.git",
]
tpu = [
    "jax[tpu]>=0.4.0",
]
gpu = [
    "jax[cuda12]>=0.4.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
jamcli = "magenta_rt_jamcli.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true